feat: 优化罗马音转换逻辑,完善备选项功能,更新文档

+++ RomajiConverter.WinUI/Models/ConvertedLine.cs
@@ -12,7 +12,7 @@ public class ConvertedLine
         Units = Array.Empty<ConvertedUnit>();
     }
 
-    public int Index { get; set; }
+    public ushort Index { get; set; }
 
     public string Chinese { get; set; }
